refactor: remove redundant entry module parameters from GraphBuildParams
The entry module is always ModuleId::ENTRY (ID 0) and the resolver already
has methods to retrieve its name and path. This refactoring eliminates the
redundant parameters from GraphBuildParams.

Changes:
- Removed entry_path and entry_module_name from GraphBuildParams
- Get entry module info from resolver using ModuleId::ENTRY
- Simplified GraphBuildParams construction

This reduces parameter passing and leverages the fact that the entry module
is a well-known constant that can be looked up when needed.